Extract test stub to its own class.
